requirements.txt
numpy==1.20.0
pandas==1.1.5
Flask==1.1.2
keras==2.4.3
matplotlib==3.3.4
opendatasets==0.1.10
Pillow==8.1.2
scikit_learn==0.24.1
seaborn==0.11.1
tensorflow==2.4.0
gunicorn==20.0.0
